remove searchview
