+++ README.md
@@ -145,9 +145,9 @@ libController.on("itemClicked", function(contextData) {
 
 ### Event 'searchTextUpdated'
 
-This event is raised when user starts typing on the search bar, and the display mode of SearchView is `list`. In this event, it should call a search algorithm from some other components, and return a list of [Search Result Items](./docs/v0.0.1/search-items.md) in JSON format to the caller.
+This event is raised when user starts typing on the search bar, and the display mode selected is `list`. In this event, it should call a search algorithm from some other components, and return a list of [Search Result Items](./docs/v0.0.1/search-items.md) in JSON format to the caller.
 
-- `searchText`: This is the value of state `searchText` in `SearchView` component, which is a string value that user has entered in the search bar.
+- `searchText`: This is the value of state `searchText` in `LibraryContainer` component, which is a string value that user has entered in the search bar.
 
  ```js
 libView.on("searchTextUpdated", function (searchText) {
